Converted performative args to be ulong

+++ inc/frame_codec.h
@@ -13,7 +13,7 @@ extern "C" {
 #endif /* __cplusplus */
 
 	typedef void* FRAME_CODEC_HANDLE;
-	typedef void(*FRAME_RECEIVED_CALLBACK)(void* content, uint32_t performative, AMQP_VALUE frame_list_value);
+	typedef void(*FRAME_RECEIVED_CALLBACK)(void* content, uint64_t performative, AMQP_VALUE frame_list_value);
 
 	extern FRAME_CODEC_HANDLE frame_codec_create(IO_HANDLE io, FRAME_RECEIVED_CALLBACK frame_received_callback, void* context, LOGGER_LOG logger_log);
 	extern void frame_codec_destroy(FRAME_CODEC_HANDLE handle);
